Changed launchfiles and added a bringup script for the simulator

+++ lolo_drivers/scripts/dvl_to_altitude.py
@@ -3,7 +3,7 @@
 # vim:fenc=utf-8
 import rospy
 from cola2_msgs.msg import DVL
-from std_msgs.msg import Float64
+from std_msgs.msg import Float32
 
 def dvl_cb(msg):
     global depth_pub
@@ -14,7 +14,7 @@ if __name__ == "__main__":
     rospy.init_node("dvl_to_altitude")
     robot_name = "lolo"
 
-    alt_pub = rospy.Publisher("/"+robot_name+"/dr/altitude", Float64, queue_size=1)
+    alt_pub = rospy.Publisher("/"+robot_name+"/dr/altitude", Float32, queue_size=1)
     dvl_sub = rospy.Subscriber("/"+robot_name+"/core/dvl", DVL, dvl_cb, queue_size=1)
 
     while not rospy.is_shutdown():
